arenasql: check not null constraint during insert
  - also make column index u8 since dont need more than 256 columns

+++ crates/arenasql/src/schema/column.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 
 use super::{DataType, SerializedCell};
 
-pub type ColumnId = u16;
+pub type ColumnId = u8;
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct Column {
